pyproject.toml
[tool.poetry]
name = "cs6065-final-proj"
version = "0.1.0"
description = "Management for CS6065 final project making a data science dashboard"
authors = ["Cat Luong <luongcn@mail.uc.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
scikit-learn = "^1.4.1.post1"
numpy = "^1.26.4"
pyodbc = "^5.1.0"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = "^2.7.0"
azure-identity = "^1.16.0"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
flask-cors = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
